[Move] logger from `src/code/common` to `src/code/platform`
