+++ src/code/platform/logger/common/abstractLoggerService.ts
