+++ src/code/platform/logger/common/consoleLoggerService.ts
